Feature: Automatic docker container build
With this commit an automatic github docker build process is established.

The docker image creation process was redefined and clear volumes were specified in order to store task and timewarrior data.

Furthermore variable declaration routines were defined in order to automatic populate the `.env` file for the use of `dotenvy` routine in `src/main.rs`.

The docker image is created in addition to the binary release on `main` branch. On every pull request, the docker image is created as well to ensure, that the docker image is still be able to build, but it is not uploaded to the package repository of `github`.