Feature: Light mode
This commit introduces the first light theme for
`taskwarrior-web` with the feature of an automatic
mode to switch between dark and light depending on
OS / Browser setting.

Variable TWK_THEME can be used in order to pre-define
a theme.

Currently only two themes are supported:
1. taskwarrior-dark
2. taskwarrior-light

By default, the theme is switching automatically.

Closes tmahmood/taskwarrior-web#8