Improve mobile usability
This commit optimizes the mobile experience and usability.
It ensures, that mobile browsers do not show the website
as it would be a desktop one, but try to scale it accordingly
to the mobile screen.

Furthermore the header is made sticky. It still kept on top
while scrolling, but can take more room if it is required due
to flex flow changes on small screens.

Relates #16